+++ src/main/java/firstApp/entity/Universite.java
@@ -15,7 +15,7 @@ import java.io.Serializable;
 public class Universite implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int idUniversite;
+    private Long idUniversite;
     private String nomUniversite;
     private String adrese;
     @OneToOne
